pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "md2docx"
version = "0.1.0"
description = "Convert Markdown + Mermaid diagrams to .docx"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1",
]

[project.urls]
Homepage = "https://github.com/unsaid-azizov/md2docx"

[project.scripts]
md2docx = "md2docx.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/md2docx"]
